alert dialog

+++ components/modal/superadmin/timekeeping/leave/viewLeave.vue
@@ -86,9 +86,26 @@
       dialog: false,
     }),
     methods: {
-      deleteLeave () {
-        this.$store.dispatch('leave/delete', this.leave.id)
-        this.dialog = false;
+      async deleteLeave () {
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!'
+        }).then(async(result) => {
+          if (result.value) {
+            await this.$store.dispatch('leave/delete', this.leave.id)
+            this.dialog = false;
+            Swal.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+          }
+        })
       }
     }
   }
