+++ components/modal/superadmin/timekeeping/overtime/viewOvertime.vue
@@ -66,8 +66,25 @@
     }),
     methods: {
       deleteOvertime () {
-        this.$store.dispatch('overtime/delete', this.overtime.id);
-        this.dialog = false;
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!'
+        }).then(async(result) => {
+          if (result.value) {
+            await this.$store.dispatch('overtime/delete', this.overtime.id);
+            this.dialog = false;
+            Swal.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+          }
+        })
       }
     }
   }
